requirements.txt
alembic==1.16.4
apscheduler==3.11.1
Flask==3.1.2
flask_cors==6.0.1
Flask_JWT_Extended==4.4.4
Flask_Migrate==4.0.0
flask_sqlalchemy==3.1.1
fpdf==1.7.2
pytest==9.0.2
python-dotenv==1.2.1
python_dateutil==2.9.0.post0
SQLAlchemy==2.0.45
Werkzeug==3.1.4
yagmail==0.15.293
